remove certain relative paths in #include, fix renderer SWAP method note to self always make sure swap method is up to date

+++ collection/graphics/Image.cpp
@@ -1,6 +1,6 @@
 
 #include "Image.h"
-#include "../io/File.h"
+#include "io/File.h"
 
 Image::Image(const std::string& filePath) {
   File file(filePath);
@@ -13,6 +13,6 @@ std::string Image::getType() const {
   return type;
 }
 
-uint32_t Image::getImageSize() const {
+int64_t Image::getImageSize() const {
   return bufSize;
 }
