+++ collection/graphics/Image.h
@@ -9,12 +9,12 @@ class Image {
  public:
   Image(const std::string& filePath);
 
-  uint32_t getImageSize() const;
+  int64_t getImageSize() const;
   std::string getType() const;
 
   std::unique_ptr<char[]> buf;
 
  private:
   std::string type;
-  uint32_t bufSize;
+  int64_t bufSize;
 };
